Adding: TestShopPopUp Tests
